etl/pyproject.toml
[tool.poetry]
name = "etl-to-ch"
version = "0.1.0"
description = ""
authors = ["Alexander Ershov <a.ershov@lancktele.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
clickhouse-driver = "^0.2.8"
more-itertools = "^10.3.0"
tenacity = "^8.5.0"
kafka-python = "^2.0.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pylint = "^2.17.4"
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
disable = "C0114,W0511, R0916, R0903"
fail-under = 8.0
extension-pkg-whitelist = "rapidjson"

[tool.pylint."MESSAGES CONTROL"]
generated-members = "df"
good-names = "df"
